#5941: move restore-local-names regression test into XmirTest
